remove keytar tests
